+++ src/redox.rs
@@ -1,6 +1,3 @@
-use core::ffi::c_void;
-use core::ptr::NonNull;
-
 use super::DisplayHandle;
 
 /// Raw display handle for the Redox operating system.
@@ -53,47 +50,5 @@ impl DisplayHandle<'static> {
     pub fn orbital() -> Self {
         // SAFETY: No data is borrowed.
         unsafe { Self::borrow_raw(OrbitalDisplayHandle::new().into()) }
-    }
-}
-
-/// Raw window handle for the Redox operating system.
-///
-/// ## Thread Safety
-///
-/// The underlying window is a [file descriptor], and most calls on the window
-/// correspond directly to non-mutating file descriptor reads and writes.
-/// Therefore this type is `Send` and `Sync`. This means that this type can be
-/// sent to and used from other threads.
-///
-/// [file descriptor]: https://github.com/redox-os/orbclient/blob/77c28e88fcb180c750175f2dcf5c7342d357ab26/src/sys/orbital.rs#L64-L65
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct OrbitalWindowHandle {
-    /// A pointer to an orbclient window.
-    // TODO(madsmtm): I think this is a file descriptor, so perhaps it should
-    // actually use `std::os::fd::RawFd`, or some sort of integer instead?
-    pub window: NonNull<c_void>,
-}
-
-unsafe impl Send for OrbitalWindowHandle {}
-unsafe impl Sync for OrbitalWindowHandle {}
-
-impl OrbitalWindowHandle {
-    /// Create a new handle to a window.
-    ///
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use core::ptr::NonNull;
-    /// # use raw_window_handle::OrbitalWindowHandle;
-    /// # type Window = ();
-    /// #
-    /// let window: NonNull<Window>;
-    /// # window = NonNull::from(&());
-    /// let mut handle = OrbitalWindowHandle::new(window.cast());
-    /// ```
-    pub fn new(window: NonNull<c_void>) -> Self {
-        Self { window }
     }
 }