+++ tests/web_handles.rs
@@ -3,7 +3,7 @@
 #![cfg(target_family = "wasm")]
 
 use core::mem::ManuallyDrop;
-use raw_window_handle::{WasmBindgenCanvasWindowHandle, WasmBindgenOffscreenCanvasWindowHandle};
+use raw_window_handle::RawWindowHandle;
 use wasm_bindgen::convert::{IntoWasmAbi, RefFromWasmAbi};
 use wasm_bindgen::JsCast;
 use web_sys::{HtmlCanvasElement, OffscreenCanvas};
@@ -23,11 +23,14 @@ fn html_canvas_element() {
 
     // Convert to the raw index and convert to the handle.
     let index = (&canvas).into_abi();
-    let handle = WasmBindgenCanvasWindowHandle::new(index as usize);
+    let handle = RawWindowHandle::new_wasm_bindgen_canvas(index as usize);
 
     // To get the canvas element back, convert the index back.
+    let RawWindowHandle::WasmBindgenCanvas { obj, .. } = handle else {
+        unreachable!()
+    };
     let other_end: ManuallyDrop<HtmlCanvasElement> =
-        unsafe { HtmlCanvasElement::ref_from_abi(handle.obj as u32) };
+        unsafe { HtmlCanvasElement::ref_from_abi(obj as u32) };
     assert_eq!(&*other_end, &canvas);
 }
 
@@ -38,11 +41,14 @@ fn offscreen_canvas() {
 
     // Convert to the raw index and convert to the handle.
     let index = (&canvas).into_abi();
-    let handle = WasmBindgenOffscreenCanvasWindowHandle::new(index as usize);
+    let handle = RawWindowHandle::new_wasm_bindgen_offscreen_canvas(index as usize);
 
     // To get the canvas element back, convert the index back.
+    let RawWindowHandle::WasmBindgenOffscreenCanvas { obj, .. } = handle else {
+        unreachable!()
+    };
     let other_end: ManuallyDrop<OffscreenCanvas> =
-        unsafe { OffscreenCanvas::ref_from_abi(handle.obj as u32) };
+        unsafe { OffscreenCanvas::ref_from_abi(obj as u32) };
     assert_eq!(&*other_end, &canvas);
 }